Rename button style files
